requirements.txt
asgiref==3.8.1
Django==5.1
gunicorn==23.0.0
mysqlclient==2.2.4
packaging==25.0
pillow==12.0.0
qrcode==8.2
six==1.17.0
sqlparse==0.5.1
tzdata==2024.1
